requirements.txt
Django==1.11
djangorestframework==3.6.2
pkg-resources==0.0.0
pytz==2018.7
